requirements.txt
Django>=5.0.2
Pillow==10.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-cors-headers==4.1.0
djangorestframework-camel-case==1.4.2
